refactor: pane import paths
