+++ apps/web/src/components/pane/index.ts
@@ -1,0 +1,1 @@
+export * from "./pane";
